+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.4.2 (not released)
 * `grandfather`: New command to add mentioned packages that ar not maintained to Grandfathered Dependencies.
+* `diff-snapshot`: Pass `--mode cabal` to generate a .cabal file for all addded/modified dependencies.
 
 ### 0.4.1
 * `add`: Tweak bound messages